pyproject.toml
[tool.poetry]
name = "cocoarag"
version = "0.1.0"
description = ""
authors = ["seeusoon <to.asmarkov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-openai = "0.2.1"
pyyaml = "^6.0.2"
python-box = "^7.2.0"
langchain-core = "^0.3.8"
langchain-postgres = "^0.0.12"
psycopg = {extras = ["binary"], version = "^3.2.3"}
langchain = "^0.3.4"
pydantic = "^2.9.2"
seaborn = "^0.13.2"
scipy = "^1.14.1"
umap-learn = "^0.5.7"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
ruff = "^0.7.4"
mypy = "^1.13.0"
pyright = "^1.1.389"
pre-commit = "^4.0.1"
black = {extras = ["jupyter"], version = "^24.10.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
